Update repos cassette
In an upcoming commit I’ll need to re-record this cassette because I’m
adding some new fields to the GraphQL query.

One new repo has been added to the dxw organization since I recorded the
current cassette. So the Repo spec needs updating to change the count of
expected repos. So I’m doing this in its own commit.

I guess this will continue to be a problem that happens each time we
need to re-record the data, when we have specs that depend on the
contents of the cassette. I will not worry about it until it becomes a
bigger problem.