Show whether repo has contributing-related files
In the table, we show whether each repo’s default branch has a
CONTRIBUTING.md and/or CODE_OF_CONDUCT.md file.

I’ve added the capybara_table gem to be able to easily interrogate the
contents of tables. There might be a not-much-more-code way of doing
this without the gem, but I didn’t want to fall into a Capybara
rabbithole right now.