Add permission check to repo
If the permissions of the dxw-rails-user are not admin or write, the service user is not able to commit code to the repo.

If the service user doesn’t have push access, it isn’t even allowed to query for its permission level, so a forbidden exception is as good as a confirmation of lack of access.

I couldn’t find a method to query for access that wouldn’t return an exception when the user has no access.