Check for ability to write to repo
When the service user can’t write to the repo, the error message doesn’t make that clear. The error message is “Not found” when attempting to create a blob.

Check for write permission and skip the repo, with a message, if the service user can’t write to it.